+++ src/controllers/Configurator.php
@@ -112,6 +112,7 @@ class Configurator extends Controller
                 $newBlockType->enabled = $settings['enabled'];
                 $newBlockType->ignorePermissions = $settings['ignorePermissions'];
                 $newBlockType->description = $settings['description'] ?? '';
+                $newBlockType->iconFilename = $settings['iconFilename'] ?? '';
                 $newBlockType->iconId = $settings['iconId'];
                 $newBlockType->minBlocks = (int)$settings['minBlocks'];
                 $newBlockType->maxBlocks = (int)$settings['maxBlocks'];