+++ src/models/BlockType.php
@@ -64,6 +64,12 @@ class BlockType extends Model implements GqlInlineFragmentInterface
     public ?string $description = null;
 
     /**
+     * @var string|null The block type's icon filename.
+     * @since 3.10.0
+     */
+    public ?string $iconFilename = null;
+
+    /**
      * @var int|null The block type's icon, as a Craft asset ID.
      * @since 3.6.0
      */
@@ -266,6 +272,17 @@ class BlockType extends Model implements GqlInlineFragmentInterface
     }
 
     /**
+     * Gets this block type's icon path, if an icon filename is set.
+     *
+     * @return string|null
+     * @since 3.10.0
+     */
+    public function getIconPath(?array $transform = null): ?string
+    {
+        return Neo::$plugin->blockTypes->getIconPath($this, $transform);
+    }
+
+    /**
      * Gets this block type's icon asset, if an icon is set.
      *
      * @return Asset|null
@@ -325,6 +342,7 @@ class BlockType extends Model implements GqlInlineFragmentInterface
             'handle' => $this->handle,
             'description' => $this->description ?? '',
             'enabled' => $this->enabled,
+            'iconFilename' => $this->iconFilename ?? '',
             'icon' => $iconData,
             'minBlocks' => (int)$this->minBlocks,
             'maxBlocks' => (int)$this->maxBlocks,